[SOLR-15450] rework the tests * separate tests that check the setup of the Feature from tests that check the ranking * separate tests for ranking from tests that observe the cache interaction * test that the delegation to the FieldValueFeature works * parameterize the enableLazyFieldLoading config setting and remove obsolete config
