+++ solr/contrib/ltr/src/test-files/solr/collection1/conf/solrconfig-ltr.xml
@@ -34,8 +34,8 @@
    initialSize="2048" autowarmCount="0" />
   <cache name="QUERY_DOC_FV" class="solr.search.CaffeineCache" size="4096"
    initialSize="2048" autowarmCount="4096" regenerator="solr.search.NoOpRegenerator" />
-  <!-- false is the default, just here to make it explizit in contrast to solrconfig-ltr-lazy-->
-  <enableLazyFieldLoading>false</enableLazyFieldLoading>
+
+  <enableLazyFieldLoading>${solr.query.enableLazyFieldLoading:false}</enableLazyFieldLoading>
  </query>
 
  <!-- add a transformer that will encode the document features in the response.
